% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFactory.R
\name{graphFactory}
\alias{graphFactory}
\title{graphFactory}
\usage{
graphFactory(graph = "std")
}
\arguments{
\item{transformation}{A keyword to select the graph of peak shape assumptions among
\itemize{
\item "updown": We add inequality constraints to the successive segment specific parameters theta_1,...,theta_k so that non-decreasing changes in these parameters are always followed by non-increasing changes;
\item "updown_negbin": "updown" with reversed constraints. Specific to the negative binomial noise model.
\item "std": The sequence of changes is not constrained.
}}
}
\value{
A function which takes as argument the peanlty \code{lambda} parameter and returns the graph of constraints.
}
\description{
A function factory that returns, on the top of the \code{gfpop} method (Runge et al. 2020), graphs which encode peak shape assumptions.
}
\examples{
print(std_g <- graphFactory(graph="std"))
}
