% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lossFactory.R
\name{lossFactory}
\alias{lossFactory}
\title{lossFactory}
\usage{
lossFactory(type = "mean")
}
\arguments{
\item{type}{A keyword to select the noise model and associated weighted loss among
\itemize{
\item "mean" (gaussian): \code{sum((data-theta)^2*weights)};
\item "poisson": \code{sum((theta-data*log(theta))*weights)};
\item "negbin" (negative binomial): \code{sum((-log(theta)-data*log(1-theta))*weights)}.
}}
}
\value{
A function which takes as arguments the observations \code{data}, associated \code{weights} and \code{theta} the distribution paramter affected by changes. The function returns the loss computed on the obsevations.
}
\description{
A function factory that returns several weighted loss forms.
}
\examples{
loss <- lossFactory(type="mean")
}
