% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformationFactory.R
\name{transformationFactory}
\alias{transformationFactory}
\title{transformationFactory}
\usage{
transformationFactory(transformation = "anscombe_poisson")
}
\arguments{
\item{transformation}{A keyword to select the type of transformation among
\itemize{
\item "raw": \code{x};
\item "anscombe_poisson": \code{sqrt(x+3/8)};
\item "anscombe_negbin": \code{sqrt(x+1/(2*phi))};
\item "delta_negbin": \code{log(2*sqrt(phi*x*(phi*x+1))+2*phi*x+1)};
\item "delta_poisson": \code{sqrt(x)};
\item "logp1": \code{log(x+1)}.
}}
}
\value{
A function which takes as arguments the observations \code{x} (and \code{phi} when needed) and returns the transformed observations.
}
\description{
A function factory that returns several useful variance stabilizing transformations for data under Poisson or negative binomial noise models.
}
\examples{
require(data.table)
counts <- as.data.table(CROCS::counts)
counts <- counts[sample.id == "McGill0036" & chunk==3,]
head(coverage <- counts$coverage)
tr <- transformationFactory(transformation="anscombe_poisson")
head(tr(x=coverage))
tr <- transformationFactory(transformation="anscombe_negbin")
head(tr(x=coverage, phi=10))
}
